feat(server): add citations database table
This commit will add citations as a unique data type. It will come with a minor refactor in the
server code which removes the simple citations as a column in the quotes table, making citations
more robust and allowing for better features: giving users a way to adjust the citations to their
unique academic needs.